Base of V3
Most updates for V3 done.

Todo:
Finish Updating README.md
Bigfix
Transfer updates to play.py
