--- Data/teams.json
@@ -1,12 +1,0 @@
-[
-    {
-        "team": "1",
-        "score": "0",
-        "members": ["billy", "bob"]
-    },
-    {
-        "team": "2",
-        "score": "0",
-        "members": []
-    }
-]
