+++ main.py
@@ -45,13 +45,31 @@ def printtable(gridfile: str) -> None:
         openfile.seek(0)
         print(openfile.read())
 
+# Prints current list of members for a single team 
+def printmembers(targetteam: int, teamlist: str | None = teamlist) -> None:
+    with open(teamlist) as openfile:
+        scoresnteams = json.load(openfile)
+        print('Members:')
+        while int(len(scoresnteams[int(targetteam)]['members'])) > 0:
+            memberslist = scoresnteams[int(targetteam)]['members']
+            print(memberslist.pop())
+        print()
+
 # Prints current lits of teams and scores
 def printteams(numofteams: int, doneteams: int | None = 1, teamlist: str | None = teamlist) -> None:
     with open(teamlist) as openfile:
         scoresnteams = json.load(openfile)
         while int(doneteams) <= int(numofteams):
-            print('Team ' + scoresnteams[doneteams-1]['team'] + ': ' + scoresnteams[doneteams-1]['score'])
-            doneteams = doneteams + 1
+            if str(scoresnteams[doneteams-1]['teamname']) == '':
+                print('Team ' + scoresnteams[doneteams-1]['team'] + ': ' + scoresnteams[doneteams-1]['score'])
+                if scoresnteams[doneteams-1]['members'] != []:
+                    printmembers(doneteams-1)
+                doneteams = doneteams + 1
+            elif str(scoresnteams[doneteams-1]['teamname']) != '':
+                print(scoresnteams[doneteams-1]['teamname'] + ': ' + scoresnteams[doneteams-1]['score'])
+                if scoresnteams[doneteams-1]['members'] != []:
+                    printmembers(doneteams-1)
+                doneteams = doneteams + 1     
 
 # Checks if a number is and integer
 def isint(number: any) -> bool:
@@ -81,13 +99,13 @@ def genscoerfile(numofteams: int | None = numofteams, teamsdone: int | None = 0,
     with open(teamlist, 'w') as openfile:
         openfile.write('[\n')
         while int((teamsdone+1)) < int(numofteams):
-            scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","score": "0","members": []}'
+            scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","teamname": "","score": "0","members": []}'
             jsondata = json.loads(scoreset)
             openfile.write(json.dumps(jsondata, indent=4))
             openfile.write(',\n')
             teamsdone = teamsdone+1
         if int((teamsdone+1)) == int(numofteams):
-            scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","score": "0","members": []}'
+            scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","teamname": "","score": "0","members": []}'
             jsondata = json.loads(scoreset)
             openfile.write(json.dumps(jsondata, indent=4))
         openfile.write('\n]')
@@ -105,6 +123,120 @@ def addscorejson(team: int, teamnumpicked: int, teamlist: str | None = teamlist,
             questions = json.load(openfile)
             nscore = int(data[int(team)-1]['score']) +  int(questions[int(teamnumpicked)-1]['points'])
         data[int(team)-1]['score'] = str(str(data[int(team)-1]['score']).replace(str(data[int(team)-1]['score']), str(nscore)))
+    with open(teamlist, 'w') as f:
+        json.dump(data, f, indent=4)
+
+# Adds members to a team
+def addmembers(numofteams: int | None = numofteams, targetteam: int | None = 0, teamsdone: int | None = 0, teamlist: str | None = teamlist, temptf: bool | None = False) -> None:
+    with open(teamlist) as f:
+        prevdata = json.load(f)
+    with open(teamlist, 'w') as openfile:
+        openfile.write('[\n')
+        while int((teamsdone+1)) < int(numofteams):
+            if teamsdone+1 != targetteam:
+                scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","score": "' + str(prevdata[teamsdone]['score']) + '","members": ' + str(str(prevdata[teamsdone]['members'])).replace("'", '"') + '}'
+                jsondata = json.loads(scoreset)
+                openfile.write(json.dumps(jsondata, indent=4))
+                openfile.write(',\n')
+                teamsdone = teamsdone+1
+            if teamsdone+1 == targetteam and int((teamsdone+1)) < int(numofteams):
+                while temptf != True:
+                    print('Who are the new members you want to add?')
+                    print('If multiple seperate names with a space')
+                    nmembers = input('(str.) ')
+                    nmembers = nmembers.replace(' ', '", "')
+                    nmembers = '["' + nmembers + '"]'
+                    print('\n' + str(nmembers) + ' Please confim that these are the names you want to add')
+                    conf = input('(y/n) ')
+                    if yesorno(conf):
+                        temptf = True
+                        cs()
+                    else: continue
+                temptf = False
+                scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","score": "' + str(prevdata[teamsdone]['score']) + '","members": ' + str(nmembers) + '}'
+                jsondata = json.loads(scoreset)
+                openfile.write(json.dumps(jsondata, indent=4))
+                openfile.write(',\n')
+                teamsdone = teamsdone+1
+        if int((teamsdone+1)) == int(numofteams) and teamsdone+1 != targetteam:
+            scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","score": "' + str(prevdata[teamsdone]['score']) + '","members": ' + str(str(prevdata[teamsdone]['members']).replace("'", '"')) + '}'
+            jsondata = json.loads(scoreset)
+            openfile.write(json.dumps(jsondata, indent=4))
+            openfile.write('\n]')
+        if teamsdone+1 == targetteam and int((teamsdone+1)) == int(numofteams):
+            while temptf != True:
+                print('Who are the new members you want to add?')
+                print('If multiple seperate names with a space')
+                nmembers = input('(str.) ')
+                nmembers = nmembers.replace(' ', '", "')
+                nmembers = '["' + nmembers + '"]'
+                print('\n' + str(nmembers) + ' Please confim that these are the names you want to add')
+                conf = input('(y/n) ')
+                if yesorno(conf):
+                    temptf = True
+                    cs()
+                else: continue
+            temptf = False
+            scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","score": "' + str(prevdata[teamsdone]['score']) + '","members": ' + str(nmembers) + '}'
+            jsondata = json.loads(scoreset)
+            openfile.write(json.dumps(jsondata, indent=4))
+            openfile.write('\n]')    
+    with open(teamlist) as f:
+        data = json.load(f)
+    with open(teamlist, 'w') as f:
+        json.dump(data, f, indent=4)
+
+# Adds team name to a team
+def addteamname(numofteams: int | None = numofteams, targetteam: int | None = 0, teamsdone: int | None = 0, teamlist: str | None = teamlist, temptf: bool | None = False) -> None:
+    with open(teamlist) as f:
+        prevdata = json.load(f)
+    with open(teamlist, 'w') as openfile:
+        openfile.write('[\n')
+        while int((teamsdone+1)) < int(numofteams):
+            if teamsdone+1 != targetteam:
+                scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","teamname": "' + str(prevdata[teamsdone]['teamname']) + '","score": "' + str(prevdata[teamsdone]['score']) + '","members": ' + str(str(prevdata[teamsdone]['members'])).replace("'", '"') + '}'
+                jsondata = json.loads(scoreset)
+                openfile.write(json.dumps(jsondata, indent=4))
+                openfile.write(',\n')
+                teamsdone = teamsdone+1
+            if teamsdone+1 == targetteam and int((teamsdone+1)) < int(numofteams):
+                while temptf != True:
+                    print('What is the new team name?')
+                    teamname = input('(str.) ')
+                    print('\n' + str(teamname) + ' - Please confim that this is the new name of the team')
+                    conf = input('(y/n) ')
+                    if yesorno(conf):
+                        temptf = True
+                        cs()
+                    else: continue
+                temptf = False
+                scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","teamname": "' + str(teamname) + '","score": "' + str(prevdata[teamsdone]['score']) + '","members": ' + str(str(prevdata[teamsdone]['members'])).replace("'", '"') + '}'
+                jsondata = json.loads(scoreset)
+                openfile.write(json.dumps(jsondata, indent=4))
+                openfile.write(',\n')
+                teamsdone = teamsdone+1
+        if int((teamsdone+1)) == int(numofteams) and teamsdone+1 != targetteam:
+            scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","teamname": "' + str(prevdata[teamsdone]['teamname']) + '","score": "' + str(prevdata[teamsdone]['score']) + '","members": ' + str(str(prevdata[teamsdone]['members']).replace("'", '"')) + '}'
+            jsondata = json.loads(scoreset)
+            openfile.write(json.dumps(jsondata, indent=4))
+            openfile.write('\n]')
+        if teamsdone+1 == targetteam and int((teamsdone+1)) == int(numofteams):
+            while temptf != True:
+                print('Who are the new members you want to add?')
+                teamname = input('(str.) ')
+                print('\n' + str(teamname) + ' - Please confim that this is the new name of the team')
+                conf = input('(y/n) ')
+                if yesorno(conf):
+                    temptf = True
+                    cs()
+                else: continue
+            temptf = False
+            scoreset = '{"team": "' + str((int(teamsdone)+1)) + '","teamname": "' + str(teamname) + '","score": "' + str(prevdata[teamsdone]['score']) + '","members": ' + str(str(prevdata[teamsdone]['members'])).replace("'", '"') + '}'
+            jsondata = json.loads(scoreset)
+            openfile.write(json.dumps(jsondata, indent=4))
+            openfile.write('\n]')    
+    with open(teamlist) as f:
+        data = json.load(f)
     with open(teamlist, 'w') as f:
         json.dump(data, f, indent=4)
 
@@ -193,7 +325,7 @@ temptf = False
 # Generate a score file for each team
 genscoerfile(numofteams)
 
-# Adds mebers to teams
+# Adds members to teams
 while temptf != True:
     print('\nWould you like to add members to teams?')
     addmem = input('(y/n) ')
@@ -205,21 +337,54 @@ while temptf != True:
                 teammemadd = input('(int.) ')
                 if teammemadd.lower() == 'exit':
                     break 
-                if isint(teammemadd) and teammemadd <= numofteams:
-                    temptf2 = True
-                elif teammemadd > numofteams:
-                    print('Value is to large, try again')
-                    temptf2 = False
+                if isint(teammemadd):
+                    if teammemadd <= numofteams:
+                        temptf2 = True
+                    else:
+                        print('Value is to large, try again')
+                        temptf2 = False   
             temptf2 = False
             cs()
             if teammemadd.lower() != 'exit':
-                addnames(teammemadd)
-        temptf = True
-        cs()
-    elif temp == False:
-        temptf = True
+                addmembers(numofteams, int(teammemadd))
+            if teammemadd.lower() == 'exit':
+                temptf2 = True
+                temptf = True
+                break 
     else:
-        temptf = False    
+        temptf = True
+temptf2 = True
+temptf = False
+
+# Addes Names to team (Team Names)
+while temptf != True:
+    print('\nWould you like to add team names to teams?')
+    addname = input('(y/n) ')
+    temp = yesorno(addname)
+    if temp:
+        while teamnameadd.lower() != 'exit':
+            while temptf2 != True:
+                print('\nWhat team are you adding members to?')
+                teamnameadd = input('(int.) ')
+                if teamnameadd.lower() == 'exit':
+                    break 
+                if isint(teamnameadd):
+                    if teamnameadd <= numofteams:
+                        temptf2 = True
+                    else:
+                        print('Value is to large, try again')
+                        temptf2 = False   
+            temptf2 = False
+            cs()
+            if teamnameadd.lower() != 'exit':
+                addteamname(numofteams, int(teamnameadd))
+            if teamnameadd.lower() == 'exit':
+                temptf2 = True
+                temptf = True
+                break 
+    else:
+        temptf = True
+temptf2 = True
 temptf = False
 
 # Figure out the starting team
